Remove validation fails halts build on development
Having the build fail because of validation fails > 0 disables Test Driven Development (because of the division of labour and the fact that testing + migration are done on different branches). The only way to commit tests up front would be to comment out the entity and this runs counter to our group decision to never have hidden (comented out) entity fails

We took a team decision to allow fails to register in the log, for full visibility, rather than hide them, and we are keeping daily tabs on the validation fails results in daily standup.