Properly cope with perform_transformations exceptions
perform_transformations() may raise EmptyDataFrame exceptions,
but the type will not be 'chunk' in this situation. Modify the
functions which get basic table data so that they recognise these
exception types and correctly set more_records.

more_records = True if:

* The current dataframe has > 0 records before filtering
* The current dataframe has 0 records after filtering

more_records = False if:

* The current dataframe has 0 records before filtering; in this case,
EmptyDataFrame is raised before perform_transformations() is called

The transformer() function doesn't care about EmptyDataFrame exceptions,
only what's in the dataframe and whether there are more records to look for.
When deciding whether to create an empty table, we only do so if
there are no more records (more_records == False) and we haven't inserted
any records at all (e.g. they were all filtered out). We also log an
error if we created the table without inserting any records, so that
we get an early warning that something might be awry with the table.