IN-472 Tasks migration scripts (#382)
* Use binary package for psycopg2

This removes the need to compile this package, which fails on
my machine anyway.

* Add tasks entity

* Include tasks transforms in local runs

* Add mapping for person_task

* Filter out tasks with status other than ACTIVE or INACTIVE

* Additional log statements to track what's happening to transform tables

* Correctly filter by status

Filter so we only have INACTIVE and ACTIVE status tasks. Then
reset the status to "Not started" for the tasks which are left
after filtering.

* Fix the conditional for deciding whether to do the db_insert for tasks

* Improve/reduce comments

* Add an include_values condition

This allows filtering to leave records whose fields match the
include_values conditions. The main use case is for tasks,
where we want tasks with a Status of ACTIVE or INACTIVE.

* Remove filtering by status from transform script

As the filter is now specified in the spreadsheet, remove
manual filtering which was applied here.

* Create table if EmptyDataFrame exception was raised during chunking

Co-authored-by: Dimitar Kokov <dimitar.kokov@digital.justice.gov.uk>

* Use binary package for psycopg2

This removes the need to compile this package, which fails on
my machine anyway.

* Add tasks entity

* Include tasks transforms in local runs

* Add mapping for person_task

* Filter out tasks with status other than ACTIVE or INACTIVE

* Additional log statements to track what's happening to transform tables

* Correctly filter by status

Filter so we only have INACTIVE and ACTIVE status tasks. Then
reset the status to "Not started" for the tasks which are left
after filtering.

* Fix the conditional for deciding whether to do the db_insert for tasks

* Improve/reduce comments

* Add an include_values condition

This allows filtering to leave records whose fields match the
include_values conditions. The main use case is for tasks,
where we want tasks with a Status of ACTIVE or INACTIVE.

* Remove filtering by status from transform script

As the filter is now specified in the spreadsheet, remove
manual filtering which was applied here.

* Create table if EmptyDataFrame exception was raised during chunking

Co-authored-by: Dimitar Kokov <dimitar.kokov@digital.justice.gov.uk>

* Remove misleading SQL

There is no need for an assignee with ID 1, so remove bogus line
which makes it look like there could be.

* Add tasks entity tables to integration

Co-authored-by: Dimitar Kokov <dimitar.kokov@digital.justice.gov.uk>