Quote column names with spaces
In the case of multiple transformations on a table based off the
same column (in this case, "End Date"), the CREATE TABLE statement
is incorrect, as table names are not wrapped in quotes. This
causes a breakage.

By wrapping all column names in quotes, we ensure that the CREATE
TABLE statement remains valid.