IN-465 Reporting date calculations (#397)
* Rework do_calculations()

Functions are not used elsewhere and are trivial, so remove
that file and associated tests.

The do_calculations() test was hard to follow, so rewrote to properly
test whether calculations are being applied.

* Put unit tests in for calculate_date field calculation

* Happy path for do_calculations/delta_date calculation

Renamed calculate_date to delta_date, as it makes it more
obvious what the calculation is doing.

* Unhappy path unit tests for do_calculations/delta_date

* Use correct format for the calculations dict

* Adjust do_calculations to cope with real data

* Move reporting date calculations into standard transformations

* Add script to run unit tests

* Ignore directories called 'build'

This is so I can create artefacts locally (e.g. coverage reports)
and not accidentally commit them to the repo all the time.

* Ignore docker log files

* Quote column names with spaces

In the case of multiple transformations on a table based off the
same column (in this case, "End Date"), the CREATE TABLE statement
is incorrect, as table names are not wrapped in quotes. This
causes a breakage.

By wrapping all column names in quotes, we ensure that the CREATE
TABLE statement remains valid.

* Add basic validation SQL for reporting/annual_report_logs

* Remove unused imports

* Remove calculate_reminderdate

As we no longer require the transform for
annual_report_logs/reportduereminderdate (we're no longer mapping
this field), remove the associated transform function.