Ensure output column keys and values have proper type
If the output_cols for a mapping only has a single column,
pandas interprets that value as a dict view, rather than a list.
Explicitly cast the keys and values for the dict to lists, so
that pandas correctly uses them to assign values to columns in
rows matching the query built from a mapping.