IN-997 Migrate report type assignments (#459)
* Add order_id to annual_report_logs migration and add report type assignment

* Add unit test for (part of) annual report type assignments

This tests the part of the transform which assigns reporttype
and type to the dataframe (i.e. the part which doesn't do any
SQL queries against the database).

* Don't use casrec orders when populating annual_report_logs

This table should just be populated from a join between the
casrec account and pat tables.

* Add more debugging when datatype mapping fails

* Ensure order_id is an Int64 column

* Cast annual_report_logs.order_id to correct datatype

* Cast annualreport_id to correct type during transform

* Return None if ord_risk_lvl is not recognised

IN-997 specifies a set of risk levels which map to particular
report types. The code previously returned OPG102 even if the
risk level was not in this set. Change so that we return None
*unless* the risk level is recognised (and there is an active order).

* Add missing annual_report_type_assignments table definition