No need to write to the table_definitions.json file lots of times
We were reading and writing the JSON file each time we read a spreadsheet.
Instead, just write it once at the end.

The other problem with the previous approach was that it loaded
the existing definitions then replaced the ones relating to each
spreadsheet in turn. However, if a table definition was removed
from the spreadsheet, it would not be removed from the table definitions
JSON. This new approach overwrites the JSON file each time the
definitions are generated.