Update warnings mapping spreadsheet
Warnings are derived from two tables. However, none of the columns
from those tables are used to create records in the target tables:
it's just the case numbers which are used, with one record being
created for each case.

Due to how is_complete was working (sort of ignoring "Yes"), the
correct columns were being included in the get basic table data query,
even though they were never actually used to set target column values.
However, when I fixed is_complete, it was necessary to remove the
table/column specifications for fields which are not in fact derived
from source columns.

Once these are removed, the SELECT to get the initial data no longer
works, as the columns used in the casrec_conditions (in the
table_defintions) are now not accessible: they are not used
in the table mapping any more, so not included in the SELECT.
To work around this, I added the columns required by
casrec_conditions to the source_table_additional_columns, as
these columns are included in the SELECT.

This then necessitated changing the casrec_conditions to use
the aliased names for these columns (e.g. c_vwm instead of VWM).