IN-1061 Map missing foreign keys for reporting tables (#491)
* Use glob to find spreadsheet files

This gets around issues where .DS_Store files turn up in the
spreadsheets directory etc.

* Clean up reporting mapping spreadsheet

* Remove unused files

Looks like a refactoring attempt, but not used and getting a bit
late for that now, so removed.

* No need to write to the table_definitions.json file lots of times

We were reading and writing the JSON file each time we read a spreadsheet.
Instead, just write it once at the end.

The other problem with the previous approach was that it loaded
the existing definitions then replaced the ones relating to each
spreadsheet in turn. However, if a table definition was removed
from the spreadsheet, it would not be removed from the table definitions
JSON. This new approach overwrites the JSON file each time the
definitions are generated.

* Reduce file count

These functions are only used from a single module and don't have
unit tests, so there's little value putting them into separate files.

* Rework mapping generation

Remove some of the duplication and unintuitive logic from
the mapping code.

Also recognise "Yes" in the is_complete column as meaning the
column has been mapped.

* Fix other erroneous mappings

* Some of the warnings tables had is_complete=Yes. This means they
  were previously excluded from validation. When I made the change
  to recognise these fields as being complete and mapped, the
  validation SQL failed, as the casrec_table was set for these
  fields in the spreadsheet, but casrec_column_name was not. By
  removing values from both of these columns in the spreadsheet,
  the default is used to construct the validation SQL.
* The finance_ledger spreadsheet had is_complete=YES for the reference
  field. This was not correctly treated as is_complete=yes, so the
  field was being ignored in the validation SQL. Once this changed so
  that it was included, the validation failed, as the reference
  field is populated with a UUID. By ignoring this during validation,
  we are still able to auto-generate the validation SQL for this table.

* Update warnings mapping spreadsheet

Warnings are derived from two tables. However, none of the columns
from those tables are used to create records in the target tables:
it's just the case numbers which are used, with one record being
created for each case.

Due to how is_complete was working (sort of ignoring "Yes"), the
correct columns were being included in the get basic table data query,
even though they were never actually used to set target column values.
However, when I fixed is_complete, it was necessary to remove the
table/column specifications for fields which are not in fact derived
from source columns.

Once these are removed, the SELECT to get the initial data no longer
works, as the columns used in the casrec_conditions (in the
table_defintions) are now not accessible: they are not used
in the table mapping any more, so not included in the SELECT.
To work around this, I added the columns required by
casrec_conditions to the source_table_additional_columns, as
these columns are included in the SELECT.

This then necessitated changing the casrec_conditions to use
the aliased names for these columns (e.g. c_vwm instead of VWM).

* Better handle error condition where casrec csv files not found

* Manually apply changes to Orders spreadsheet