Add transform from casrec pat -> Sirius annual_report_logs
Pending reports are in the casrec pat table, not in account.
Add a transform to extract pat records and make annual_report_logs
from them, using the pat record's latest reporting period
+ 1 day as the reportingperiodstartdate.

Note that this has to be done manually rather than with a
mapping spreadsheet, as a mapping already exists for
annual_report_logs. See IN-1096 for more explanation.