Don't use simple transform for tasks duedate
It's not possible to use a column from the casrec source table
as both the source column for a simple mapping and as an input
to a simple transform. This is because do_simple_mapping()
renames the columns, rather than adding a new column with the
target name. Consequently, if you attempt to refer to a column
used in a simple mapping from the specification for a simple
transform, the column no longer exists and it fails.