IN-1095 Set tasks due date to casrec.sup_activity Target or Start Date (#560)
* Add unit tests for coalesce()

Mirrors the SQL COALESCE function, setting a column to the first
non-empty value in a list of columns.

* Implement coalesce() standard transformation

* Use coalesce() transform to set tasks duedate

Use casrec sup_activity ActStart or Start Date for duedate on Sirius task.

* Don't use simple transform for tasks duedate

It's not possible to use a column from the casrec source table
as both the source column for a simple mapping and as an input
to a simple transform. This is because do_simple_mapping()
renames the columns, rather than adding a new column with the
target name. Consequently, if you attempt to refer to a column
used in a simple mapping from the specification for a simple
transform, the column no longer exists and it fails.

* Remove duplicated code

Lots of the lines in this file are redundant, so do some
reflection to instead call the transform functions by name
as required.

* Manually use coalesce() in the tasks transform function

As we can't apply the coalesce() function as a simple transform,
manually apply it to the dataframe in the tasks transform.

* Better handling of null column values

Rather than only checking for None and "", use pandas.notnull()
and "" instead. This ensures that values like pandas.NaT and
numpy.nan (as well as None) are recognised correctly.

* Fix tasks validation

* Merge changes to spreadsheet after rebase

Although the changes to the mapping JSON could be merged
automatically, spreadsheet has to be manually updated so
that it corresponds.