Add special cases reindexing step
The current primary/fk reindexing isn't flexible enough
to cope with updating an arbitrary property inside a JSON
column, based on changes to foreign keys on dependent tables.

This is precisely what is needed to update the JSON in
scheduled_events for reporting, where we have references to
annual_report_logs.id and annual_report_logs.client_id.

Add a custom reindex function which applies the necessary
changes to scheduled_events.