IN-1177 Remove activedate from migration
Slight complexity in that duedate was calculated in part from activedate, which was itself based on Start Date.

With activedate gone, we are free to map duedate to Start Date directly in the mappings, which is then set as its default

Then the default behaviour is unchanged: If Target Date is present, then the duedate value is overwritten (using that coalesce function).

Validation coalesce functionality hasn’t changed