IN1201 - Changes for SW-5461
Sirius post-migration task SW-5461 created many duplicate deputies in order to save deputy addresses. Duplicates were marked as ‘Legacy’ and clients & orders were RE-MAPPED to new deputies.

The impact of the above work on this ticket is that when we came to update ~2000 pro/pa rows in the order_deputy table, the rows were no longer there, because the orders are now being mapped to a replacement (de-duped) deputy.

Working with Dimitar it has been possible to recreate the re-map from two files outputted by the remap script - seen in the sql as:
- IN1201_split_deputies_final.csv
- IN1201_new_deputies_final.csv

From these files we create a lookup table ({pmf_schema}.changed_deputies) that we can use in the update script.

The update table thus changes to make two joins to the order_deputy table, one using the original deputy id, and one using the updated deputy id, and coalesces the results)

NOTE the remapping performed by Sirius fully removed the row from order_deputy - consequently there is no overlap between od and od2 in the SQL - where od is found, od2 is NULL, and vicaversa